test/pyproject.toml
[project]
name = "test"
version = "0.1.0"
description = "Treebeard test scripts"
requires-python = ">=3.12"
dependencies = [
    "dnspython>=2.8.0",
]
